+++ carbonmark-api/test/plugins/rate-limit.test.ts
@@ -2,8 +2,7 @@ import { FastifyInstance } from "fastify";
 import nock from "nock";
 import { GRAPH_URLS } from "../../src/graphql/codegen.constants";
 import { build } from "../helper";
-import { CATEGORIES } from "../routes/routes.mock";
-import { DEV_URL } from "../test.constants";
+import { CATEGORIES, DEV_URL } from "../test.constants";
 
 describe("Rate Limiter", () => {
   let app: FastifyInstance;
