+++ carbonmark-api/test/routes/purchases/get.test.ts
@@ -5,8 +5,7 @@ import { aPurchase } from "../../../src/.generated/mocks/marketplace.mocks";
 import { GRAPH_URLS } from "../../../src/graphql/codegen.constants";
 import { Purchase } from "../../../src/models/Purchase.model";
 import { build } from "../../helper";
-import { DEV_URL } from "../../test.constants";
-import { ERROR } from "../routes.mock";
+import { DEV_URL, ERROR } from "../../test.constants";
 
 const mockPurchase = aPurchase();
 
@@ -37,7 +36,6 @@ describe("GET /purchases/:id", () => {
   let fastify: FastifyInstance;
 
   // Setup the server
-  afterEach(async () => await fastify.close());
   beforeEach(async () => {
     fastify = await build();
   });
