--- carbonmark-api/test/routes/routes.mock.ts
@@ -1,29 +1,0 @@
-export const ERROR = {
-  message: "Graph error occurred",
-  locations: [{ line: 6, column: 7 }],
-  path: ["graph", 1],
-};
-
-export const CATEGORIES = [
-  {
-    id: "Blue Carbon",
-  },
-  {
-    id: "Forestry",
-  },
-  {
-    id: "A Random Category",
-  },
-];
-
-export const COUNTRIES = [
-  {
-    id: "Canada",
-  },
-  {
-    id: "Mozambique",
-  },
-  {
-    id: "A Random Country",
-  },
-];
