+++ carbonmark-api/test/test.constants.ts
@@ -1,2 +1,32 @@
 export const DEV_URL = "http://localhost:3003";
 export const MOCK_ADDRESS = "0x0123456789012345678901234567890123456789";
+
+export const ERROR = {
+  message: "Graph error occurred",
+  locations: [{ line: 6, column: 7 }],
+  path: ["graph", 1],
+};
+
+export const CATEGORIES = [
+  {
+    id: "Blue Carbon",
+  },
+  {
+    id: "Forestry",
+  },
+  {
+    id: "A Random Category",
+  },
+];
+
+export const COUNTRIES = [
+  {
+    id: "Canada",
+  },
+  {
+    id: "Mozambique",
+  },
+  {
+    id: "A Random Country",
+  },
+];
